requirements.txt
discord.py
python-dotenv
requests
asyncio
pytz
datetime
numpy
pymongo[srv]